refactor(client): consolidate and rename the login/set-password page files
Two tidy-ups for the client-side auth pages:

- Merge the login page's two stylesheets into one (dropping a dead
  .login-sso rule), so each page is one <name>.css like set-password.
- Rename login_page.* -> login.* and set_password_page.* -> set_password.*
  (component, stylesheet and spec). The "_page" suffix only existed to avoid
  colliding with the old login.ts / set_password.ts entry stubs, which were
  removed when these screens moved to the client; the collision is gone.

Updates the index.ts dynamic-import dispatch accordingly.
