+++ apps/client/src/login.tsx
@@ -1,5 +1,5 @@
 import "./setup.css";
-import "./login_page.css";
+import "./login.css";
 
 import { LOCALE_IDS } from "@triliumnext/commons";
 import { render } from "preact";
